+++ src/main/java/act/route/RouterAdmin.java
@@ -21,12 +21,8 @@ package act.route;
  */
 
 import act.app.App;
-import act.cli.CliContext;
-import act.cli.Command;
-import act.cli.Optional;
-import act.cli.Required;
+import act.cli.*;
 import act.cli.tree.FilteredTreeNode;
-import act.cli.tree.TreeNode;
 import act.cli.tree.TreeNodeFilter;
 import act.util.PropertySpec;
 import org.osgl.http.H;
@@ -58,38 +54,8 @@ public class RouterAdmin {
             @Optional("specify route filter") String q
     ) {
         final Router router = S.blank(name) ? app.router() : app.router(name);
-        if (S.notBlank(q)) {
-            if (q.contains(".") || q.contains("[") || q.contains("*")) {
-                // already regex
-            } else {
-                // make it a regex
-                q = ".*" + q + ".*";
-            }
-        }
         if (tree) {
-            TreeNode root = new TreeNode() {
-
-                @Override
-                public String id() {
-                    return "root";
-                }
-
-                @Override
-                public String label() {
-                    return "Router";
-                }
-
-                @Override
-                public List<TreeNode> children() {
-                    List<TreeNode> l = new ArrayList<>();
-                    l.add(router._GET);
-                    l.add(router._POST);
-                    l.add(router._PUT);
-                    l.add(router._DEL);
-                    return l;
-                }
-            };
-            return S.blank(q) ? root : new FilteredTreeNode(root, TreeNodeFilter.Common.pathMatches(q));
+            return S.blank(q) ? router : new FilteredTreeNode(router, TreeNodeFilter.Common.pathMatches(q));
         } else {
             return routeInfoList(name, q);
         }